+++ domain/src/main/java/com/example/domain/model/Product.kt
@@ -1,3 +1,3 @@
 package com.example.domain.model
 
-data class Product(val id: Int, val name: String, val price: Int, val imageUrl: String)
+data class Product(val id: Long, val name: String, val price: Int, val imageUrl: String)
