+++ domain/src/main/java/com/example/domain/repository/ProductRepository.kt
@@ -5,5 +5,5 @@ import com.example.domain.model.Product
 interface ProductRepository {
     fun getAll(): List<Product>
     fun getNext(count: Int): List<Product>
-    fun findById(id: Int): Product
+    fun findById(id: Long): Product
 }
